scraping/requirements.txt
Scrapy==2.6.1
python-dateutil==2.8.2
Twisted==22.4.0